Create Workflow in Form Assembly
I have a two page form. I want the first page to be submitted, and create a lead in SFDC. Then the second page knows which lead was created on the first page so it can write more information to that lead. This is accomplished in Form Assembly using two forms, what they call "Workflow" and the "Object Alias". The user isn't aware that they submitted two forms. It just looks like page 1 and page 2.