Create FormAssembly Redirect and stay inside VF
Form Assembly form 1 is embedded in a VF page so it looks like you never left Salesforce. 
When that form is submitted, a lead is created, and the form redirects to a second FA form inside visualforce.
The FA redirect link is https://mysalesorcedomainblahblah/partners/apex/formassemblypage2?leadid=%%SFA_LEAD%% (which is the id for the lead created by the first form).  
So that's the link to the visualforce page plus "?leadid=%%SFA_LEAD%%"
The code is for the Controller for the 2nd form. 
I'm not using the formassembly salesforce prefill connector, but prefilling through the URL.